+++ include/private_cube.h
@@ -4,6 +4,6 @@
 void draw_cube_precalc(void);
 void kociemba_precalc(void);
 
-void cube_set_gap(unsigned int degree);
+void cube_set_gap(float degree);
 
 #endif /* !HEADER_RUBIK_DRAW_CUBE */
